remove redundant middleware for admin checks via query, passed from the req

+++ _features/text-channels/textChannels.model.js
@@ -51,14 +51,15 @@ const getTextChannelById = async (groupId, textChannelId) => {
 };
 
 // UPDATE A TEXT CHANNEL
-const updateTextChannel = async (groupId, textChannelId, newName) => {
+const updateTextChannel = async (groupId, textChannelId, newName, is_admin, userId) => {
   const result = await pool.query(
     `UPDATE text_channels 
      SET name = $1
      WHERE group_id = $2 
      AND id = $3
+     and (created_by = $4 or $5 = true)
      RETURNING *`,
-    [newName, groupId, textChannelId]
+    [newName, groupId, textChannelId, userId, is_admin]
   );
 
   if (result.rows.length === 0) {
@@ -69,13 +70,14 @@ const updateTextChannel = async (groupId, textChannelId, newName) => {
 };
 
 // DELETE A TEXT CHANNEL
-const deleteTextChannel = async (groupId, textChannelId) => {
+const deleteTextChannel = async (groupId, textChannelId, is_admin, userId) => {
   const result = await pool.query(
     `DELETE FROM text_channels 
      WHERE group_id = $1 
-     AND id = $2 
+     AND id = $2
+     AND (created_by = $3 OR $4 = true)
      RETURNING *`,
-    [groupId, textChannelId]
+    [groupId, textChannelId, userId, is_admin]
   );
 
   if (result.rows.length === 0) {
